# examples/kratos/kratos-example/api/helloworld/v1/user_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v3.21.12
// source: helloworld/v1/user.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationUserAdminOnly = "/helloworld.v1.User/AdminOnly"
const OperationUserEditUser = "/helloworld.v1.User/EditUser"
const OperationUserGetUserInfo = "/helloworld.v1.User/GetUserInfo"
const OperationUserLogin = "/helloworld.v1.User/Login"
const OperationUserLogout = "/helloworld.v1.User/Logout"
const OperationUserPublicInfo = "/helloworld.v1.User/PublicInfo"

type UserHTTPServer interface {
	// AdminOnly 管理员接口（需要admin角色）
	AdminOnly(context.Context, *AdminRequest) (*AdminReply, error)
	// EditUser 需要特定权限的接口（需要user.edit权限）
	EditUser(context.Context, *EditUserRequest) (*EditUserReply, error)
	// GetUserInfo 获取用户信息（需要登录）
	GetUserInfo(context.Context, *GetUserInfoRequest) (*GetUserInfoReply, error)
	// Login 用户登录
	Login(context.Context, *LoginRequest) (*LoginReply, error)
	// Logout 用户登出
	Logout(context.Context, *LogoutRequest) (*LogoutReply, error)
	// PublicInfo 公开接口（无需登录）
	PublicInfo(context.Context, *PublicInfoRequest) (*PublicInfoReply, error)
}

func RegisterUserHTTPServer(s *http.Server, srv UserHTTPServer) {
	r := s.Route("/")
	r.POST("/api/login", _User_Login0_HTTP_Handler(srv))
	r.POST("/api/logout", _User_Logout0_HTTP_Handler(srv))
	r.GET("/api/user/info", _User_GetUserInfo0_HTTP_Handler(srv))
	r.GET("/api/admin/dashboard", _User_AdminOnly0_HTTP_Handler(srv))
	r.POST("/api/user/edit", _User_EditUser0_HTTP_Handler(srv))
	r.GET("/api/public/info", _User_PublicInfo0_HTTP_Handler(srv))
}

func _User_Login0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserLogin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Login(ctx, req.(*LoginRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginReply)
		return ctx.Result(200, reply)
	}
}

func _User_Logout0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LogoutRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserLogout)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Logout(ctx, req.(*LogoutRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LogoutReply)
		return ctx.Result(200, reply)
	}
}

func _User_GetUserInfo0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetUserInfoRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserGetUserInfo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetUserInfo(ctx, req.(*GetUserInfoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetUserInfoReply)
		return ctx.Result(200, reply)
	}
}

func _User_AdminOnly0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AdminRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserAdminOnly)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AdminOnly(ctx, req.(*AdminRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AdminReply)
		return ctx.Result(200, reply)
	}
}

func _User_EditUser0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in EditUserRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserEditUser)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.EditUser(ctx, req.(*EditUserRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*EditUserReply)
		return ctx.Result(200, reply)
	}
}

func _User_PublicInfo0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PublicInfoRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserPublicInfo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.PublicInfo(ctx, req.(*PublicInfoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PublicInfoReply)
		return ctx.Result(200, reply)
	}
}

type UserHTTPClient interface {
	// AdminOnly 管理员接口（需要admin角色）
	AdminOnly(ctx context.Context, req *AdminRequest, opts ...http.CallOption) (rsp *AdminReply, err error)
	// EditUser 需要特定权限的接口（需要user.edit权限）
	EditUser(ctx context.Context, req *EditUserRequest, opts ...http.CallOption) (rsp *EditUserReply, err error)
	// GetUserInfo 获取用户信息（需要登录）
	GetUserInfo(ctx context.Context, req *GetUserInfoRequest, opts ...http.CallOption) (rsp *GetUserInfoReply, err error)
	// Login 用户登录
	Login(ctx context.Context, req *LoginRequest, opts ...http.CallOption) (rsp *LoginReply, err error)
	// Logout 用户登出
	Logout(ctx context.Context, req *LogoutRequest, opts ...http.CallOption) (rsp *LogoutReply, err error)
	// PublicInfo 公开接口（无需登录）
	PublicInfo(ctx context.Context, req *PublicInfoRequest, opts ...http.CallOption) (rsp *PublicInfoReply, err error)
}

type UserHTTPClientImpl struct {
	cc *http.Client
}

func NewUserHTTPClient(client *http.Client) UserHTTPClient {
	return &UserHTTPClientImpl{client}
}

// AdminOnly 管理员接口（需要admin角色）
func (c *UserHTTPClientImpl) AdminOnly(ctx context.Context, in *AdminRequest, opts ...http.CallOption) (*AdminReply, error) {
	var out AdminReply
	pattern := "/api/admin/dashboard"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserAdminOnly))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// EditUser 需要特定权限的接口（需要user.edit权限）
func (c *UserHTTPClientImpl) EditUser(ctx context.Context, in *EditUserRequest, opts ...http.CallOption) (*EditUserReply, error) {
	var out EditUserReply
	pattern := "/api/user/edit"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserEditUser))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetUserInfo 获取用户信息（需要登录）
func (c *UserHTTPClientImpl) GetUserInfo(ctx context.Context, in *GetUserInfoRequest, opts ...http.CallOption) (*GetUserInfoReply, error) {
	var out GetUserInfoReply
	pattern := "/api/user/info"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserGetUserInfo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Login 用户登录
func (c *UserHTTPClientImpl) Login(ctx context.Context, in *LoginRequest, opts ...http.CallOption) (*LoginReply, error) {
	var out LoginReply
	pattern := "/api/login"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserLogin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Logout 用户登出
func (c *UserHTTPClientImpl) Logout(ctx context.Context, in *LogoutRequest, opts ...http.CallOption) (*LogoutReply, error) {
	var out LogoutReply
	pattern := "/api/logout"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserLogout))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// PublicInfo 公开接口（无需登录）
func (c *UserHTTPClientImpl) PublicInfo(ctx context.Context, in *PublicInfoRequest, opts ...http.CallOption) (*PublicInfoReply, error) {
	var out PublicInfoReply
	pattern := "/api/public/info"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserPublicInfo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}