# examples/kratos/kratos-example/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/login:
        post:
            tags:
                - User
            description: 用户登录
            operationId: User_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/helloworld.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.LoginReply'
...
    /api/user/edit:
        post:
            tags:
                - User
            description: 需要特定权限的接口（需要user.edit权限）
            operationId: User_EditUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/helloworld.v1.EditUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.EditUserReply'
...
components:
    schemas:
...
        helloworld.v1.EditUserReply:
            type: object
            properties:
                message:
                    type: string
            description: 编辑用户响应
        helloworld.v1.EditUserRequest:
            type: object
            properties:
                userId:
                    type: string
                newUsername:
                    type: string
            description: 编辑用户请求
...
        helloworld.v1.LoginReply:
            type: object
            properties:
                token:
                    type: string
                message:
                    type: string
                userId:
                    type: string
            description: 登录响应
        helloworld.v1.LoginRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
            description: 登录请求
...